requirements.txt
# API
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2

# LLM — OpenRouter uses the OpenAI SDK
openai>=1.54.0

# Embeddings + Vector Search
sentence-transformers==3.1.1
faiss-cpu>=1.9.0
numpy==1.26.4

# Scraping (only needed for scraper.py)
playwright>=1.48.0
beautifulsoup4==4.12.3

# Utilities
python-dotenv==1.0.1
requests==2.32.3